+++ purchases/src/main/kotlin/com/revenuecat/purchases/UpgradeInfo.kt
@@ -1,0 +1,8 @@
+package com.revenuecat.purchases
+
+import com.android.billingclient.api.BillingFlowParams
+
+data class UpgradeInfo(
+    val oldSku: String,
+    @BillingFlowParams.ProrationMode val prorationMode: Int? = null
+)